+++ samples/Chart/33_Chart_create_multiple_charts.php
@@ -158,7 +158,7 @@ $chart2 = new Chart(
     $legend2, // legend
     $plotArea2, // plotArea
     true, // plotVisibleOnly
-    0, // displayBlanksAs
+    DataSeries::EMPTY_AS_GAP, // displayBlanksAs
     null, // xAxisLabel
     $yAxisLabel2 // yAxisLabel
 );